Uncoupled user confirmation from the main loop Confirmation routines are provided as a module object Will allow for better "testability"
