+++ lib/main-loop.js
@@ -3,7 +3,6 @@ const async = require('async');
 const logging = require('./logging');
 const progress = require('./progress');
 const csvReader = require('./input/csv-file-reader');
-const {confirm_Async} = require('./input/confirm-migration-params');
 const cloudinary = require('cloudinary').v2;
 
 /* ℹ️ 👇 Modules intended to be customized */
@@ -22,7 +21,7 @@ let payloadLog = null;
 //
 // Migration flow implementation
 //
-async function loopOverCsvInput_Async (payloadModule, cliParams) {
+async function loopOverCsvInput_Async (payloadModule, cliParams, confirmationRoutinesModule) {
     // Retrieving required command line parameters passed for the invocation
     const inputCsvFilePath = cliParams.fromCsvFile;
     const maxConcurrentUploads = cliParams.maxConcurrentUploads;
@@ -37,14 +36,23 @@ async function loopOverCsvInput_Async (payloadModule, cliParams) {
 
     await ensureCloudinaryConfigOrExit_Async();
 
-    const migrationOptions = {
+    const operationOptions = {
         dest_cloud             : cloudinary.config().cloud_name,
         parameters             : cliParams,
     }
 
-    await confirmMigrationOptionsOrExit_Async(migrationOptions);
+    try {
+        await confirmationRoutinesModule.confirmOperationOptionsOrExit_Async(operationOptions);
+    } catch (err) {
+        const msg = 'Migration parameters not confirmed. Terminating';
+        console.error(`🛑 ${msg}`);
+        scriptLog.fatal(operationOptions, msg);
+        // Allowing bunyan to "catch up" on writing the log file: https://github.com/trentm/node-bunyan/issues/37
+        await new Promise(resolve => setTimeout(resolve, 500));
+        process.exit(1);
+    }
 
-    scriptLog.info(migrationOptions, 'Migration parameters confirmed. Starting migration routine');
+    scriptLog.info(operationOptions, 'Migration parameters confirmed. Starting migration routine');
     const stats = {
         concurrent: 0,
         attempted: 0,
@@ -102,31 +110,6 @@ async function ensureCloudinaryConfigOrExit_Async() {
         const message = 'Cloudinary config is not initialized. Please set CLOUDINARY_URL environment variable (explicitly or via .env file).'
         console.error(`🛑 ${message}`);
         scriptLog.info(message);
-        // Allowing bunyan to "catch up" on writing the log file: https://github.com/trentm/node-bunyan/issues/37
-        await new Promise(resolve => setTimeout(resolve, 500));
-        process.exit(1);
-    }
-}
-
-/**
- * Prompts user to confirm parameters for the migration operation.
- * If confirmed - returns. Otherwise - exits the script.
- * 
- * @param {Object} migrationOptions 
- * @param {string} migrationOptions.dest_cloud    -  Destination Cloudinary environment (sub-account)
- * @param {string} migrationOptions.from_csv_file -  Path to the CSV file with the migration input
- */
-async function confirmMigrationOptionsOrExit_Async(migrationOptions) {
-    const migrationPrompt = 
-`❗️WARNING: This script will perform asset migration with the following parameters:
-${JSON.stringify(migrationOptions, null, 2)}
-
-Are you sure you want to proceed?`;
-    const promptConfirmed = await confirm_Async(migrationPrompt);
-    if (!promptConfirmed) {
-        const msg = 'Migration parameters not confirmed. Terminating';
-        console.error(`🛑 ${msg}`);
-        scriptLog.fatal(migrationOptions, msg);
         // Allowing bunyan to "catch up" on writing the log file: https://github.com/trentm/node-bunyan/issues/37
         await new Promise(resolve => setTimeout(resolve, 500));
         process.exit(1);
